+++ .github/workflows/delphi.yml
@@ -35,7 +35,7 @@ jobs:
         platform:
           - ubuntu-latest
         toolchain:
-          - 1.41.0 # MSRV
+          - 1.44.0 # MSRV
           - stable
     runs-on: ${{ matrix.platform }}
     steps:
@@ -79,7 +79,7 @@ jobs:
       - uses: actions-rs/toolchain@v1
         with:
           profile: minimal
-          toolchain: 1.41.0 # MSRV
+          toolchain: 1.44.0 # MSRV
           components: clippy
       - run: cargo clippy --all --all-features -- -D warnings
