+++ Cargo.toml
@@ -31,7 +31,9 @@ once_cell = "1"
 serde = { version = "1", features = ["serde_derive"] }
 serde_json = "1"
 sha2 = "0.9"
-stdtx = "0.2.3"
+stdtx = "0.3"
+tendermint = "0.16"
+tendermint-rpc = "0.16"
 thiserror = "1"
 tokio = { version = "0.2", features = ["rt-threaded", "macros"] }
 warp = "0.2"
